requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.1.1
Django==4.1.2
frozenlist==1.3.1
future==0.18.2
gunicorn==20.1.0
idna==2.10
line-bot-sdk==2.3.0
multidict==6.0.2
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
tzdata==2022.5
urllib3==1.26.4
wincertstore==0.2
yarl==1.8.1
